saving the data on object
