+++ index.html
@@ -34,6 +34,6 @@
     <div id="task-list"></div>
 
     <!-- js -->
-    <script type="module" src="/assets/js/add-task.js"></script>
+    <script type="module" src="/assets/js/main.js"></script>
   </body>
 </html>
